Whitelist the fellowship APIs
