+++ backend/src/main/java/org/sefglobal/core/fellowship/model/Certificate.java
@@ -1,8 +1,9 @@
-package org.sefglobal.core.model;
+package org.sefglobal.core.fellowship.model;
 
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.Setter;
+import org.sefglobal.core.model.AuditModel;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
@@ -12,7 +13,7 @@ import java.util.Date;
 @Table(name = "certificate")
 @Getter
 @Setter
-public class Certificate extends AuditModel{
+public class Certificate extends AuditModel {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
