+++ backend/src/main/java/org/sefglobal/core/fellowship/repository/CertificateRepository.java
@@ -1,6 +1,6 @@
-package org.sefglobal.core.repository;
+package org.sefglobal.core.fellowship.repository;
 
-import org.sefglobal.core.model.Certificate;
+import org.sefglobal.core.fellowship.model.Certificate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CertificateRepository extends JpaRepository<Certificate, Integer> {
